+++ src/Command/TorrentList.php
@@ -4,8 +4,6 @@ namespace Popstas\Transmission\Console\Command;
 
 use Martial\Transmission\API\Argument\Torrent;
 use Popstas\Transmission\Console\Helpers\TorrentListUtils;
-use Popstas\Transmission\Console\Helpers\TorrentUtils;
-use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -19,9 +17,9 @@ class TorrentList extends Command
             ->setName('torrent-list')
             ->setAliases(['tl'])
             ->setDescription('List torrents')
-            ->addOption('sort', null, InputOption::VALUE_OPTIONAL, 'Sort by column number', 4)
-            ->addOption('age', null, InputOption::VALUE_OPTIONAL, 'Sort by torrent age, ex. \'>1 <5\'')
             ->addOption('name', null, InputOption::VALUE_OPTIONAL, 'Sort by torrent name (regexp)')
+            ->addOption('age', null, InputOption::VALUE_OPTIONAL, 'Sort by torrent age, ex. \'>1 <5\'')
+            ->addOption('sort', null, InputOption::VALUE_OPTIONAL, 'Sort by column number', 4)
             ->addOption('limit', null, InputOption::VALUE_OPTIONAL, 'Limit torrent list')
             ->setHelp(<<<EOT
 The <info>torrent-list</info> list torrents.